Allow usage in NodeJS with the embedded dagre module
The dagre module detects the module exports in the NodeJS environment and
does not declare the dagre object globally.  The code in nomnoml.layouter.js
fails accessing the global dagre variable.  Let's declate self = {} to make
dagre put its object to it, when running outside of the browser.

A better solution would be to require the dagre module like lodash and
limiting the supported versions to 0.4.x, which have the needed interface.